Fix ssh agent forwarding with docker desktop (#21)
SSH agent forwarding was not working properly for users with recent
versions of Docker Desktop on macOS (and possibly other platforms). This
is likely due to restrictions on how files on the host are exposed to
the VM in which docker is running.

Docker Desktop however exposes the user's ssh agent, if any, at the
following path `/run/host-services/ssh-auth.sock`. If we detect the user
is using Docker Desktop, we bind that instead of relying on
`$SSH_AUTH_SOCK`.

Fixes #20 

Drive by: Update dependencies and pre-commit hooks