Ignore unknown keys in bitbucket-pipelines.yml (#29)
Since this tool doesn't (and can't) support all features offered by
Bitbucket pipelines, some keys that could be in a user's pipelines files
are not mapped by this tool. This would lead to an error because unknown
keys were forbidden. Users could comment out those keys to use the tool,
but this can be somewhat inconvenient and could lead to committing
unwanted changes by accident.

To alleviate this pain point, unknown keys will now be ignored by
default. An effort will still be made to parse keys related to
unsupported features, such as oidc, and emit a warning to at least
inform the user that something is unsupported.

Fixes #28 

Drive-by: Update dependencies