chore: Update docker service to latest version (#34)
Update the docker service to version 25.0.3, to be in sync with the
version used by Bitbucket Pipelines. Strangely, the version of the cli
they inject if needed is not the same, it is actually 20.10.24. The
injected docker cli was changed to match this version. The sha256 sum of
the injected docker cli was validated to ensure it matches the one they
inject.

The test for docker buildkit was also updated to actually require
buildkit, for added assurance.

Drive-by: Remove un-necessary ignored ruff rules. Those rules have been
removed from ruff.