feat: Add support for block style artifact declaration (#54)
Bitbucket Pipelines supports artifacts in two forms, the first of which
is already supported:
* A list of file paths
* Block of new-line separated key-value pairs

This adds support for the second form. In addition to specifying a list
of paths, the second form also has a download flag, which controls if
the previously stored artifacts should be downloaded to the container.
This is now also supported.

Fixes #51 

Reference:
https://support.atlassian.com/bitbucket-cloud/docs/step-options/#Caches-and-Artifacts