Add missing pipes env vars (#57)
When running pipes, bitbucket injects a bunch of environment variables to the docker container it runs, including most BITBUCKET_* variables and `CI`. it also injects some docker variables. This reproduces that behaviour.

Some limitations are still in place. The `BITBUCKET_DOCKER_HOST_INTERNAL` is not yet set, so it can't be propagated to the pipe, which means that docker-in-docker won't work for pipes yet.