fix: Ensure artifacts are loaded/saved even if container doesn't have `tar` (#61)
In the unlikely, but realistic case, where the image chosen to run a
pipeline doesn't have `tar` pre-installed, it would be impossible to
load and save artifacts. Even worse, it would fail silently.

To fix this, we use a dedicated container, guaranteed to include `tar`,
that mounts the target container's volumes to extract and restore
artifacts.