Add support for stages (#69)
Add support for pipelines that contain stages. By default, all the
stages of the selected pipeline will be run, in order of declaration.
The user will be prompted to run stages that have a manual trigger.

It is possible to specify which stages should run, with the `--stage`
parameter. Multiple stages can be specified by using the parameter
multiple times.

Conditions on stages are ignored, with a warning log.